+++ apply-labels/src/jobs/__init__.py
@@ -1,0 +1,1 @@
+from .jobs import *
